fix: add fix for common path function when formatting files at directory root and tests
