+++ package.json
@@ -16,7 +16,7 @@
     "format": "npm run build && node dist/cli.js 'src/**/*.ts'",
     "release": "release-it",
     "watch": "tsc --watch",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "vitest"
   },
   "keywords": [],
   "author": {
@@ -42,7 +42,8 @@
     "eslint-plugin-unicorn": "^48.0.1",
     "prettier": "^3.0.0",
     "release-it": "^16.1.3",
-    "typescript": "^5.1.6"
+    "typescript": "^5.1.6",
+    "vitest": "^0.34.3"
   },
   "peerDependencies": {
     "eslint": "8.x",
