+++ vite.config.ts
@@ -1,0 +1,8 @@
+// vite.config.ts
+import { defineConfig } from "vitest/config"
+
+export default defineConfig({
+  test: {
+    globals: true
+  }
+})
